myproject/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django==4.2.11",
    "ipython==8.22.2",
    "django-debug-toolbar==4.3.0",
    "pillow==10.2.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "django-cors-headers==4.7.0",
    "ultralytics==8.3.82",
    "drf-yasg[validation]",
    "coreapi",
    "django-jsonform==2.22.0",
    "gunicorn==22.0.0",
]